Implemented a simple Health Tracking Dashboard with Flask. Followed a procedure of implementing simple examples for each feature and after adding the exampled featured case in the app.
Flask was used for the app development along with Jinja for the templates and SQLAlchemy for the database manipulation and CRUD operations. Finally, the Chart.js library was used for graphical representation of the user's Health Data.

The user is welcomed in the Homepage and can navigate to add their Health Data per date in a Postgres Database that's running on Vercel. Finally, their Health Data results are presented in a dashboard template page where they can view graphs of their Health Data History.